requirements.txt
# Core libraries
numpy
torch
nltk
tqdm

# Bioinformatics
biopython

# NLP and retrieval
rank_bm25
sentence-transformers
